refactor: Make _get_country_list return a list of countries regardless of number of found countries
This is what the naming of the method suggests it is supposed to do and without it, typechecking fails or requires painful workarounds. .

Code that uses a "get_countries_" method and expects a single Country object as reurn type will break. Only the get_country_by_country_code has been adapted to return the first element of the list of countries that _get_country_list returns.